Shared Driver
Implemented Shared driver for tests so now the entire browser won't close and open, rather it will just delete cookies and reload the page.
Imported Spring to do this, but right now it is implemented via package, not with Spring. I need to study more on how to get it to work with Spring.